moving samples to ExampleXXX format for better inclusion in documentation
